+++ belgif-rest-problem-it/belgif-rest-problem-it-common/src/main/resources/openapi.yaml
