+++ belgif-rest-problem-it/belgif-rest-problem-openapi-validation-spring-boot-2-it/.gitignore
@@ -1,0 +1,1 @@
+**/openapi.yaml
