+++ belgif-rest-problem-it/belgif-rest-problem-openapi-validation-spring-boot-2-it/src/main/java/io/github/belgif/rest/problem/Application.java
