+++ belgif-rest-problem-it/belgif-rest-problem-openapi-validation-spring-boot-2-it/src/test/java/io/github/belgif/rest/problem/RestProblemOpenapiValidationSpringBoot2IT.java
@@ -1,0 +1,20 @@
+package io.github.belgif.rest.problem;
+
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.server.LocalServerPort;
+
+import io.restassured.RestAssured;
+import io.restassured.specification.RequestSpecification;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+class RestProblemOpenapiValidationSpringBoot2IT extends AbstractOpenApiValidationSpringBootIT {
+
+    @LocalServerPort
+    private int port;
+
+    @Override
+    protected RequestSpecification getSpec() {
+        return RestAssured.with().baseUri("http://localhost").port(port).basePath("/spring");
+    }
+
+}
