+++ belgif-rest-problem-it/belgif-rest-problem-openapi-validation-spring-boot-3-it/src/main/resources/application.properties
