+++ belgif-rest-problem-it/pom.xml
@@ -37,7 +37,8 @@
   <modules>
     <module>belgif-rest-problem-spring-boot-2-it</module>
     <module>belgif-rest-problem-spring-boot-3-it</module>
-    <module>belgif-rest-problem-openapi-validation-spring-it</module>
+    <module>belgif-rest-problem-openapi-validation-spring-boot-2-it</module>
+    <module>belgif-rest-problem-openapi-validation-spring-boot-3-it</module>
     <module>belgif-rest-problem-java-ee-it</module>
     <module>belgif-rest-problem-jakarta-ee-it</module>
     <module>belgif-rest-problem-it-common</module>
